pyproject.toml
[tool.poetry]
name = "atemplates"
version = "0.1.0"
description = ""
authors = ["Nickolsky <romb.intu@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Flask = "^2.0.1"
Flask-Login = "^0.5.0"
Werkzeug = "^2.0.1"
pycryptodome = "^3.10.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
